feat: -- 新增 SqlToy.

+++ 2.x/spring-boot-sqltoy/src/main/java/com/lmaye/sqltoy/dto/SysUserDTO.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 /**
  * @author lmayZhou
@@ -22,13 +21,13 @@ public class SysUserDTO implements Serializable {
      * 主键ID
      */
     @Schema(name = "id", description = "主键ID")
-    private BigInteger id;
+    private Long id;
 
     /**
      * 部门ID
      */
     @Schema(name = "deptId", description = "部门ID", nullable = true)
-    private BigInteger deptId;
+    private Long deptId;
 
     /**
      * 客户端ID
